Update README.md
This commit introduces an updated list of valuable resources pertaining to AI, machine learning, and specifically Large Language Models (LLM). It provides a diverse set of links ranging from research papers, blogs, to GitHub repositories and tool documentation. Each entry comes with a brief description in Markdown to offer a quick understanding of the resource. This update will enrich our README, providing users with a comprehensive view of recent advancements and practical tools in the field